+++ lib/sidekiq/mem/warden.rb
@@ -13,15 +13,40 @@ module Sidekiq
       include Sidekiq::ServerMiddleware if defined?(Sidekiq::ServerMiddleware)
 
       MUTEX = Mutex.new
+      DEFAULT_MAX_RSS_MB = 8192
+      DEFAULT_GRACE_TIME_SECONDS = 300
+      DEFAULT_SHUTDOWN_WAIT_SECONDS = 30
+      DEFAULT_KILL_SIGNAL = "SIGKILL"
+      DEFAULT_GC = true
 
       def initialize(options = {})
-        @max_rss = options.fetch(:max_rss, 2048)
-        @grace_time = options.fetch(:grace_time, 2500)
-        @shutdown_wait = options.fetch(:shutdown_wait, 30)
-        @kill_signal = options.fetch(:kill_signal, "SIGKILL")
-        @gc = options.fetch(:gc, true)
+        @max_rss = options.fetch(:max_rss, DEFAULT_MAX_RSS_MB)
+        @grace_time = options.fetch(:grace_time, DEFAULT_GRACE_TIME_SECONDS)
+        @shutdown_wait = options.fetch(:shutdown_wait, DEFAULT_SHUTDOWN_WAIT_SECONDS)
+        @kill_signal = options.fetch(:kill_signal, DEFAULT_KILL_SIGNAL)
+        @gc = options.fetch(:gc, DEFAULT_GC)
         @skip_shutdown = options.fetch(:skip_shutdown_if, proc { false })
         @on_shutdown = options.fetch(:on_shutdown, nil)
+      end
+
+      def self.options_from_env(env = ENV)
+        {
+          max_rss: env.fetch("SIDEKIQ_MEM_WARDEN_LIMIT_MB", DEFAULT_MAX_RSS_MB).to_i,
+          grace_time: env.fetch("SIDEKIQ_MEM_WARDEN_GRACE_TIME", DEFAULT_GRACE_TIME_SECONDS).to_i,
+          shutdown_wait: env.fetch("SIDEKIQ_MEM_WARDEN_SHUTDOWN_WAIT", DEFAULT_SHUTDOWN_WAIT_SECONDS).to_i,
+          kill_signal: env.fetch("SIDEKIQ_MEM_WARDEN_KILL_SIGNAL", DEFAULT_KILL_SIGNAL),
+          gc: truthy?(env.fetch("SIDEKIQ_MEM_WARDEN_GC", DEFAULT_GC.to_s))
+        }
+      end
+
+      def self.install!(config, env: ENV, logger: Sidekiq.logger, **overrides)
+        options = options_from_env(env).merge(overrides)
+        config.server_middleware do |chain|
+          chain.add(self, **options)
+        end
+
+        logger&.warn(install_log(options))
+        options
       end
 
       def call(worker, job, queue)
@@ -56,7 +81,13 @@ module Sidekiq
 
       def request_shutdown
         Thread.new do
-          shutdown if MUTEX.try_lock
+          next unless MUTEX.try_lock
+
+          begin
+            shutdown
+          ensure
+            MUTEX.unlock if MUTEX.locked?
+          end
         end
       end
 
@@ -101,6 +132,8 @@ module Sidekiq
       def sidekiq_process
         Sidekiq::ProcessSet.new.find do |process|
           process["identity"] == identity
+        end || Sidekiq::ProcessSet.new.find do |process|
+          process["pid"].to_i == ::Process.pid
         end || raise("No sidekiq worker with identity #{identity} found")
       end
 
@@ -114,6 +147,20 @@ module Sidekiq
 
       def warn(msg)
         Sidekiq.logger.warn(msg)
+      end
+
+      def self.truthy?(value)
+        return value if value == true || value == false
+
+        !%w[0 false no off].include?(value.to_s.strip.downcase)
+      end
+
+      def self.install_log(options)
+        "[sidekiq-mem-warden] installed limit=#{options[:max_rss]}MB " \
+          "grace=#{options[:grace_time]}s " \
+          "shutdown_wait=#{options[:shutdown_wait]}s " \
+          "kill_signal=#{options[:kill_signal]} " \
+          "gc=#{options[:gc]}"
       end
     end
   end